refactor(common): unify observability with app factory and middleware
Introduce unified observability infrastructure with app factory pattern and
standardized middleware across all services.

## Major Changes

### New Infrastructure
- **App Factory** (services/common/app_factory.py): Standardized FastAPI app
  creation with automatic observability setup, instrumentation, and middleware
  registration
- **Observability Middleware** (services/common/middleware.py): Unified
  middleware for correlation ID propagation, request/response logging, and
  timing metrics
- **HTTP Headers** (services/common/http_headers.py): Standardized HTTP
  header utilities

### Service Refactoring
- Updated all service apps to use create_service_app() factory:
  - audio, bark, discord, flan, guardrails, orchestrator, stt, testing
- Standardized observability setup across services
- Improved tracing integration with enhanced service instrumentation

### Common Utilities
- Enhanced tracing.py with improved observability manager and service setup
- Improved structured_logging.py with better context handling
- Updated http_client.py, resilient_http.py, and optimized_http.py with
  standardized patterns
- Enhanced configuration system in services/common/config/

### Documentation
- Comprehensive observability documentation in docs/operations/observability.md
- Updated service deep dives for orchestrator and STT
- Enhanced shared utilities documentation
- Updated configuration catalog and development setup guides

### Infrastructure
- Major Makefile restructuring and optimization (364 lines changed)
- Enhanced build scripts with improved Docker image management
- Added new utility scripts: docker-ghcr-auth.sh, download-models.sh
- Updated service URL documentation

### Fixes
- Fix missing Awaitable import in middleware.py
- Fix Makefile linting violation in restart target

## Impact
- All services now use consistent observability patterns
- Automatic correlation ID propagation across service boundaries
- Standardized request/response logging with timing metrics
- Reduced boilerplate in service initialization
- Improved maintainability and debugging capabilities

## Validation
✅ All linting checks pass
✅ All unit and component tests pass (275 unit, 68 component)
✅ Type checking passes (mypy)
✅ Test coverage: 21.91% (above 10% requirement)

+++ services/common/config/presets.py
@@ -218,7 +218,7 @@ def get_service_preset(service_name: str) -> dict[str, Any]:
                 "model": "medium.en",
                 "model_path": "/app/models",
                 "device": "cpu",
-                "compute_type": "float16",
+                "compute_type": "int8",
                 "audio_service_url": "http://audio:9100",
                 "audio_service_timeout": 50.0,
                 "enable_enhancement": True,
